client/requirements.txt
# HTTP 请求
requests>=2.31.0

# WebSocket（实时语音）
websockets>=12.0

# JSON Schema 校验
jsonschema>=4.17.0

# 环境变量管理
python-dotenv>=1.0.0

# 日志增强（可选）
colorlog>=6.7.0
